Add calibrations_only keyword to red_setupworkdir_*
The same keyword is added to red_setupworkdir_trippel and
red_setupworkdir_slitjaw for full compatibility with the other
two subroutines, red_setupworkdir_chromis and red_setupwordir_crisp.

At the moment, in this two subroutines this keyword is dummy and does
nothin. In the future, some extra functionality might be added.