Add a for-loop over root directores
A for-loop over root directores stored in root_dir from a list of
found_dir is added to search for data split between different
storages.  For each roor_dir a doit.pro/config.txt file pair is
generated.  Each such doit.pro script must be executed separately
one by one as the current pipeline does not support several root
directores in the reduction class.

In most of the cases, only one root directory is found an only one
pair of script/config file is generated.  In rare cases several root
directories might be available and this is considered to be an
exception that need a careful interactive treatment by the user.