Add suffix to config/script file names
If several root directories are found, then for each of them the
corresponding config and script files must be generated.  To
distinguish these files a suffix made of the root directory index
must be added right before the extension.

If only one  root directory is found (this is the default and the
most likely situation) then the suffix is empty and the script and
the config file names are unchanged.