Add logging in red_getcalibdata
If a calibration data collection is partially done, it makes sense
to restart from the unprocessed part and skip what has been done
before.  To achieve this, put all the processed lines from the
doit*.pro script to a separate done*.log file.  When the collection
is restarted on the same data set, use stored lines from the
done*.log file to choose which lines are done and which need to be
processed.